pyproject.toml
[tool.poetry]
name = "pt.backend"
version = "0.1.0"
description = ""
authors = ["Trung Nguyen <nttrung143@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.0"
PyJWT = "^2.1.0"
coverage = "^5.5"
celery = "^5.1.2"
redis = "^3.5.3"
requests = "^2.26.0"
boto3 = "^1.20.24"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-storages = "^1.13.2"
django-extensions = "^3.2.1"
drf-yasg = "^1.21.5"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
ipykernel = "^6.2.0"
pandas = "^1.3.2"
matplotlib = "^3.4.3"
flake8 = "^4.0.1"
black = "^21.12b0"
pre-commit = "^2.17.0"
pydotplus = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
